app/requirements/base.txt
django
django-rest-framework
django-rest-swagger
django-rest-auth
celery
mysqlclient
redis